Creating file RegisterBits which contains the bit definitions in enum classes

+++ include/snowfox/hal/riscv64/FE310/DigitalInPin.h
@@ -54,7 +54,7 @@ public:
                         volatile uint32_t       * gpio_iof_en,
                         volatile uint32_t       * gpio_input_val,
                         volatile uint32_t       * gpio_pue,
-                        uint8_t           const   in_pin_number);
+                        uint32_t          const   in_pin_number);
   virtual ~DigitalInPin();
 
 
@@ -71,12 +71,12 @@ private:
 
   volatile uint32_t       * _gpio_input_val,
                           * _gpio_pue;
-           uint8_t  const   _in_pin_number;
+           uint32_t const   _in_pin_number;
 
   static void setGpioPinAsInput(volatile uint32_t       * gpio_input_en,
                                 volatile uint32_t       * gpio_output_en,
                                 volatile uint32_t       * gpio_iof_en,
-                                         uint8_t  const   in_pin_number);
+                                         uint32_t const   in_pin_number);
 
 };
